+++ src/decrypt/crypto.rs
@@ -1,7 +1,8 @@
 use crate::global::parameters::{
     BenchMode, CipherType, HashMode, OutputFile,
 };
-use crate::global::{DecryptStreamCiphers, BLOCK_SIZE, SALT_LEN};
+use crate::global::{BLOCK_SIZE, SALT_LEN};
+use crate::global::crypto::DecryptStreamCiphers;
 use aead::stream::DecryptorLE31;
 use aead::{Aead, NewAead};
 use aes_gcm::{Aes256Gcm, Nonce};
