+++ src/global/crypto.rs
@@ -1,0 +1,60 @@
+use aead::{
+    stream::{DecryptorLE31, EncryptorLE31},
+    Payload,
+};
+use aes_gcm::Aes256Gcm;
+use chacha20poly1305::XChaCha20Poly1305;
+
+pub enum EncryptStreamCiphers {
+    AesGcm(Box<EncryptorLE31<Aes256Gcm>>),
+    XChaCha(Box<EncryptorLE31<XChaCha20Poly1305>>),
+}
+
+pub enum DecryptStreamCiphers {
+    AesGcm(Box<DecryptorLE31<Aes256Gcm>>),
+    XChaCha(Box<DecryptorLE31<XChaCha20Poly1305>>),
+}
+
+impl EncryptStreamCiphers {
+    pub fn encrypt_next<'msg, 'aad>(
+        &mut self,
+        payload: impl Into<Payload<'msg, 'aad>>,
+    ) -> aead::Result<Vec<u8>> {
+        match self {
+            EncryptStreamCiphers::AesGcm(s) => s.encrypt_next(payload),
+            EncryptStreamCiphers::XChaCha(s) => s.encrypt_next(payload),
+        }
+    }
+
+    pub fn encrypt_last<'msg, 'aad>(
+        self,
+        payload: impl Into<Payload<'msg, 'aad>>,
+    ) -> aead::Result<Vec<u8>> {
+        match self {
+            EncryptStreamCiphers::AesGcm(s) => s.encrypt_last(payload),
+            EncryptStreamCiphers::XChaCha(s) => s.encrypt_last(payload),
+        }
+    }
+}
+
+impl DecryptStreamCiphers {
+    pub fn decrypt_next<'msg, 'aad>(
+        &mut self,
+        payload: impl Into<Payload<'msg, 'aad>>,
+    ) -> aead::Result<Vec<u8>> {
+        match self {
+            DecryptStreamCiphers::AesGcm(s) => s.decrypt_next(payload),
+            DecryptStreamCiphers::XChaCha(s) => s.decrypt_next(payload),
+        }
+    }
+
+    pub fn decrypt_last<'msg, 'aad>(
+        self,
+        payload: impl Into<Payload<'msg, 'aad>>,
+    ) -> aead::Result<Vec<u8>> {
+        match self {
+            DecryptStreamCiphers::AesGcm(s) => s.decrypt_last(payload),
+            DecryptStreamCiphers::XChaCha(s) => s.decrypt_last(payload),
+        }
+    }
+}
